+++ pydidas/core/generic_params/generic_params_settings.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -16,12 +16,12 @@
 # along with Pydidas. If not, see <http://www.gnu.org/licenses/>.
 
 """
-The generic_params_isettings module holds all the required data to create generic
+The generic_params_settings module holds all the required data to create generic
 Parameters for the global pydidas settings
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -132,7 +132,7 @@ GENERIC_PARAMS_SETTINGS = {
         "choices": ["Process in GUI", "Command line"],
         "unit": "",
         "allow_None": False,
-        "tooltip": ("Specify how the processing shall be performed."),
+        "tooltip": "Specify how the processing shall be performed.",
     },
     "plugin_path": {
         "type": str,
@@ -181,6 +181,27 @@ GENERIC_PARAMS_SETTINGS = {
         "tooltip": (
             "The fraction of pixels with low values which will be ignored when "
             "cropping the histogram for 2d plots. "
+        ),
+    },
+    "cmap_name": {
+        "type": str,
+        "default": "Gray",
+        "name": "Default colormap",
+        "choices": None,
+        "allow_None": False,
+        "unit": "",
+        "tooltip": "The default colormap used in pydidas plots.",
+    },
+    "cmap_nan_color": {
+        "type": str,
+        "default": "#9AFEFF",
+        "name": "Color for invalid data / no data",
+        "choices": None,
+        "allow_None": False,
+        "unit": "",
+        "tooltip": (
+            "The RGB color used to fill missing or invalid data points. Invalid data "
+            "points are labeled with np.NaN values."
         ),
     },
 }